DAY-023: Move `GridStack` to its own file
