+++ Days/ViewsAndModifiers/ViewsAndModifiers/ContentView.swift
@@ -37,24 +37,6 @@ struct ContentView: View {
     }
 }
 
-struct GridStack<Content: View>: View {
-    let rows: Int
-    let columns: Int
-    @ViewBuilder let content: (Int, Int) -> Content
-
-    var body: some View {
-        VStack {
-            ForEach(0 ..< rows, id: \.self) { row in
-                HStack {
-                    ForEach(0 ..< columns, id: \.self) { column in
-                        content(row, column)
-                    }
-                }
-            }
-        }
-    }
-}
-
 #Preview {
     ContentView()
 }
