+++ Days/ViewsAndModifiers/ViewsAndModifiers/GridStack.swift
@@ -1,0 +1,24 @@
+//
+// GridStack.swift
+// Last modified by 'Ángel Vázquez' in 2024.
+//
+
+import SwiftUI
+
+struct GridStack<Content: View>: View {
+    let rows: Int
+    let columns: Int
+    @ViewBuilder let content: (Int, Int) -> Content
+
+    var body: some View {
+        VStack {
+            ForEach(0 ..< rows, id: \.self) { row in
+                HStack {
+                    ForEach(0 ..< columns, id: \.self) { column in
+                        content(row, column)
+                    }
+                }
+            }
+        }
+    }
+}
